cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devexforge-cli"
version = "0.1.0"
description = "DevExForge developer self-service CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
devex = "devex.main:cli"

[tool.setuptools.packages.find]
include = ["devex*"]
